database/seeder/pyproject.toml
[project]
name = "seeder"
version = "0.0.0"
description = "Project seeder. Fill database with mock products"
readme = "README.md"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "faker>=40.4.0",
    "PyYAML"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["seeder*"]
